Parse lipid name.

parseLipidName reads the provided lipid name and returns structural information as a named vector. Will return a vector with the "Grammar" element set to "NOT_PARSEABLE" if none of the parsers was able to parse the provided name successfully. If any error was raised, returns an empty vector.

parseLipidName(lipidName)

Arguments

lipidName
The lipid name to parse.

Value

Named vector with details of the lipid, empty vector otherwise.

Examples

parseLipidName("PC 32:1")

#>                 Normalized Name                   Original Name 
#>                       "PC 32:1"                       "PC 32:1" 
#>                         Grammar             Lipid Maps Category 
#>                        "Goslin"                            "GP" 
#>           Lipid Maps Main Class           Functional Class Abbr 
#> "Glycerophosphocholines [GP01]"                          "[PC]" 
#>       Functional Class Synonyms                           Level 
#>                          "[PC]"                       "SPECIES" 
#>                         Total C                        Total OH 
#>                            "32"                             "0" 
#>                        Total DB                    FA1 Position 
#>                             "1"                            "NA" 
#>                           FA1 C                          FA1 OH 
#>                            "NA"                            "NA" 
#>                          FA1 DB                   FA1 Bond Type 
#>                            "NA"                            "NA" 
#>                    FA2 Position                           FA2 C 
#>                            "NA"                            "NA" 
#>                          FA2 OH                          FA2 DB 
#>                            "NA"                            "NA" 
#>                   FA2 Bond Type                    LCB Position 
#>                            "NA"                            "NA" 
#>                           LCB C                          LCB OH 
#>                            "NA"                            "NA" 
#>                          LCB DB                   LCB Bond Type 
#>                            "NA"                            "NA" 
#>                    FA3 Position                           FA3 C 
#>                            "NA"                            "NA" 
#>                          FA3 OH                          FA3 DB 
#>                            "NA"                            "NA" 
#>                   FA3 Bond Type                    FA4 Position 
#>                            "NA"                            "NA" 
#>                           FA4 C                          FA4 OH 
#>                            "NA"                            "NA" 
#>                          FA4 DB                   FA4 Bond Type 
#>                            "NA"                            "NA" 

parseLipidName("LPC 34:1")

#>                 Normalized Name                   Original Name 
#>                      "LPC 34:1"                      "LPC 34:1" 
#>                         Grammar             Lipid Maps Category 
#>                        "Goslin"                            "GP" 
#>           Lipid Maps Main Class           Functional Class Abbr 
#> "Glycerophosphocholines [GP01]"                         "[LPC]" 
#>       Functional Class Synonyms                           Level 
#>                 "[LPC, LysoPC]"         "STRUCTURAL_SUBSPECIES" 
#>                         Total C                        Total OH 
#>                            "34"                             "0" 
#>                        Total DB                    FA1 Position 
#>                             "1"                             "1" 
#>                           FA1 C                          FA1 OH 
#>                            "34"                             "0" 
#>                          FA1 DB                   FA1 Bond Type 
#>                             "1"                         "ESTER" 
#>                    FA2 Position                           FA2 C 
#>                            "NA"                            "NA" 
#>                          FA2 OH                          FA2 DB 
#>                            "NA"                            "NA" 
#>                   FA2 Bond Type                    LCB Position 
#>                            "NA"                            "NA" 
#>                           LCB C                          LCB OH 
#>                            "NA"                            "NA" 
#>                          LCB DB                   LCB Bond Type 
#>                            "NA"                            "NA" 
#>                    FA3 Position                           FA3 C 
#>                            "NA"                            "NA" 
#>                          FA3 OH                          FA3 DB 
#>                            "NA"                            "NA" 
#>                   FA3 Bond Type                    FA4 Position 
#>                            "NA"                            "NA" 
#>                           FA4 C                          FA4 OH 
#>                            "NA"                            "NA" 
#>                          FA4 DB                   FA4 Bond Type 
#>                            "NA"                            "NA" 

parseLipidName("TG(18:1_18:0_16:1)")

#>            Normalized Name              Original Name 
#>       "TAG 18:1-18:0-16:1"       "TG(18:1_18:0_16:1)" 
#>                    Grammar        Lipid Maps Category 
#>                "LipidMaps"                       "GL" 
#>      Lipid Maps Main Class      Functional Class Abbr 
#> "Triradylglycerols [GL03]"                     "[TG]" 
#>  Functional Class Synonyms                      Level 
#>                "[TAG, TG]"     "MOLECULAR_SUBSPECIES" 
#>                    Total C                   Total OH 
#>                       "52"                        "0" 
#>                   Total DB               FA1 Position 
#>                        "2"                       "-1" 
#>                      FA1 C                     FA1 OH 
#>                       "18"                        "0" 
#>                     FA1 DB              FA1 Bond Type 
#>                        "1"                    "ESTER" 
#>               FA2 Position                      FA2 C 
#>                       "-1"                       "18" 
#>                     FA2 OH                     FA2 DB 
#>                        "0"                        "0" 
#>              FA2 Bond Type               LCB Position 
#>                    "ESTER"                       "NA" 
#>                      LCB C                     LCB OH 
#>                       "NA"                       "NA" 
#>                     LCB DB              LCB Bond Type 
#>                       "NA"                       "NA" 
#>               FA3 Position                      FA3 C 
#>                       "-1"                       "16" 
#>                     FA3 OH                     FA3 DB 
#>                        "0"                        "1" 
#>              FA3 Bond Type               FA4 Position 
#>                    "ESTER"                       "NA" 
#>                      FA4 C                     FA4 OH 
#>                       "NA"                       "NA" 
#>                     FA4 DB              FA4 Bond Type 
#>                       "NA"                       "NA"