Parse lipid name with a specific grammar.

parseLipidName reads the provided lipid name based on grammar and returns structural information as a named vector. Will return a vector with the "Grammar" element set to "NOT_PARSEABLE" if none of the parsers was able to parse the provided name with the provided grammar successfully. If any error was raised, returns an empty vector.

parseLipidNameWithGrammar(lipidName, grammar)

Arguments

lipidName
The lipid name to parse.

grammar
The grammar to use. One of "Goslin", "GoslinFragments", "SwissLipids", "LipidMaps", "HMDB"

Value

Named vector with details of the lipid, empty vector otherwise.

Examples

parseLipidNameWithGrammar("PC 32:1", "Goslin")

#>                 Normalized Name                   Original Name 
#>                       "PC 32:1"                       "PC 32:1" 
#>                         Grammar             Lipid Maps Category 
#>                        "Goslin"                            "GP" 
#>           Lipid Maps Main Class                    Species Name 
#> "Glycerophosphocholines [GP01]"                       "PC 32:1" 
#>           Functional Class Abbr       Functional Class Synonyms 
#>                          "[PC]"                          "[PC]" 
#>                           Level                         Total C 
#>                       "SPECIES"                            "32" 
#>                        Total OH                        Total DB 
#>                             "0"                             "1" 
#>                    FA1 Position                           FA1 C 
#>                            "NA"                            "NA" 
#>                          FA1 OH                          FA1 DB 
#>                            "NA"                            "NA" 
#>                   FA1 Bond Type                    FA2 Position 
#>                            "NA"                            "NA" 
#>                           FA2 C                          FA2 OH 
#>                            "NA"                            "NA" 
#>                          FA2 DB                   FA2 Bond Type 
#>                            "NA"                            "NA" 
#>                    LCB Position                           LCB C 
#>                            "NA"                            "NA" 
#>                          LCB OH                          LCB DB 
#>                            "NA"                            "NA" 
#>                   LCB Bond Type                    FA3 Position 
#>                            "NA"                            "NA" 
#>                           FA3 C                          FA3 OH 
#>                            "NA"                            "NA" 
#>                          FA3 DB                   FA3 Bond Type 
#>                            "NA"                            "NA" 
#>                    FA4 Position                           FA4 C 
#>                            "NA"                            "NA" 
#>                          FA4 OH                          FA4 DB 
#>                            "NA"                            "NA" 
#>                   FA4 Bond Type 
#>                            "NA" 

parseLipidNameWithGrammar("LPC(34:1)", "SwissLipids")

#>                           Normalized Name 
#>                                "LPC 34:1" 
#>                             Original Name 
#>                               "LPC(34:1)" 
#>                                   Grammar 
#>                             "SwissLipids" 
#>                       Lipid Maps Category 
#>                                      "GP" 
#>                     Lipid Maps Main Class 
#> "Monoacylglycerophosphocholines [GP0105]" 
#>                              Species Name 
#>                                "LPC 34:1" 
#>                     Functional Class Abbr 
#>                                   "[LPC]" 
#>                 Functional Class Synonyms 
#>                           "[LPC, LysoPC]" 
#>                                     Level 
#>                                 "SPECIES" 
#>                                   Total C 
#>                                      "34" 
#>                                  Total OH 
#>                                       "0" 
#>                                  Total DB 
#>                                       "1" 
#>                              FA1 Position 
#>                                      "NA" 
#>                                     FA1 C 
#>                                      "NA" 
#>                                    FA1 OH 
#>                                      "NA" 
#>                                    FA1 DB 
#>                                      "NA" 
#>                             FA1 Bond Type 
#>                                      "NA" 
#>                              FA2 Position 
#>                                      "NA" 
#>                                     FA2 C 
#>                                      "NA" 
#>                                    FA2 OH 
#>                                      "NA" 
#>                                    FA2 DB 
#>                                      "NA" 
#>                             FA2 Bond Type 
#>                                      "NA" 
#>                              LCB Position 
#>                                      "NA" 
#>                                     LCB C 
#>                                      "NA" 
#>                                    LCB OH 
#>                                      "NA" 
#>                                    LCB DB 
#>                                      "NA" 
#>                             LCB Bond Type 
#>                                      "NA" 
#>                              FA3 Position 
#>                                      "NA" 
#>                                     FA3 C 
#>                                      "NA" 
#>                                    FA3 OH 
#>                                      "NA" 
#>                                    FA3 DB 
#>                                      "NA" 
#>                             FA3 Bond Type 
#>                                      "NA" 
#>                              FA4 Position 
#>                                      "NA" 
#>                                     FA4 C 
#>                                      "NA" 
#>                                    FA4 OH 
#>                                      "NA" 
#>                                    FA4 DB 
#>                                      "NA" 
#>                             FA4 Bond Type 
#>                                      "NA" 

parseLipidNameWithGrammar("TG(18:1_18:0_16:1)", "LipidMaps")

#>             Normalized Name               Original Name 
#>        "TAG 18:1-18:0-16:1"        "TG(18:1_18:0_16:1)" 
#>                     Grammar         Lipid Maps Category 
#>                 "LipidMaps"                        "GL" 
#>       Lipid Maps Main Class                Species Name 
#> "Triacylglycerols [GL0301]"                  "TAG 52:2" 
#>       Functional Class Abbr   Functional Class Synonyms 
#>                      "[TG]"                 "[TAG, TG]" 
#>                       Level                     Total C 
#>      "MOLECULAR_SUBSPECIES"                        "52" 
#>                    Total OH                    Total DB 
#>                         "0"                         "2" 
#>                FA1 Position                       FA1 C 
#>                         "0"                        "18" 
#>                      FA1 OH                      FA1 DB 
#>                         "0"                         "1" 
#>               FA1 Bond Type                FA2 Position 
#>                     "ESTER"                         "0" 
#>                       FA2 C                      FA2 OH 
#>                        "18"                         "0" 
#>                      FA2 DB               FA2 Bond Type 
#>                         "0"                     "ESTER" 
#>                LCB Position                       LCB C 
#>                        "NA"                        "NA" 
#>                      LCB OH                      LCB DB 
#>                        "NA"                        "NA" 
#>               LCB Bond Type                FA3 Position 
#>                        "NA"                         "0" 
#>                       FA3 C                      FA3 OH 
#>                        "16"                         "0" 
#>                      FA3 DB               FA3 Bond Type 
#>                         "1"                     "ESTER" 
#>                FA4 Position                       FA4 C 
#>                        "NA"                        "NA" 
#>                      FA4 OH                      FA4 DB 
#>                        "NA"                        "NA" 
#>               FA4 Bond Type 
#>                        "NA"