Parse multiple lipid names with the provided grammar and return a data frame with the results.

parseLipidNamesWithGrammar reads the provided lipid names vector and parses it against the given grammar. It returns structural information as a data frame. Will return a cell with the "Grammar" column set to "NOT_PARSEABLE" if none of the parsers was able to parse the provided name successfully. If any error was raised, returns an empty data frame.

parseLipidNamesWithGrammar(lipidNames, grammar)

Arguments

lipidNames
The vector of lipid names to parse.

grammar
The grammar to use. One of "Goslin", "GoslinFragments", "SwissLipids", "LipidMaps", "HMDB".

Value

Data frame where each row reports the parsing result of each element in lipidNames.

Examples

parseLipidNamesWithGrammar(c("PC 32:1","LPC 34:1","TG(18:1_18:0_16:1)"), "Goslin")

#> Warning: Could not find matching parser for grammar 'Goslin'! Valid ones are 'Goslin', 'GoslinFragments', 'SwissLipids', 'LipidMaps', 'HMDB'

#>                    Normalized Name      Original Name       Grammar
#> PC 32:1                    PC 32:1            PC 32:1        Goslin
#> LPC 34:1                  LPC 34:1           LPC 34:1        Goslin
#> TG(18:1_18:0_16:1)              NA TG(18:1_18:0_16:1) NOT_PARSEABLE
#>                    Lipid Maps Category                   Lipid Maps Main Class
#> PC 32:1                             GP           Glycerophosphocholines [GP01]
#> LPC 34:1                            GP Monoacylglycerophosphocholines [GP0105]
#> TG(18:1_18:0_16:1)                  NA                                      NA
#>                    Species Name Functional Class Abbr Functional Class Synonyms
#> PC 32:1                 PC 32:1                  [PC]                      [PC]
#> LPC 34:1               LPC 34:1                 [LPC]             [LPC, LysoPC]
#> TG(18:1_18:0_16:1)           NA                    NA                        NA
#>                                    Level Total C Total OH Total DB FA1 Position
#> PC 32:1                          SPECIES      32        0        1           NA
#> LPC 34:1           STRUCTURAL_SUBSPECIES      34        0        1            1
#> TG(18:1_18:0_16:1)                    NA      NA       NA       NA           NA
#>                    FA1 C FA1 OH FA1 DB FA1 Bond Type FA2 Position FA2 C FA2 OH
#> PC 32:1               NA     NA     NA            NA           NA    NA     NA
#> LPC 34:1              34      0      1         ESTER           NA    NA     NA
#> TG(18:1_18:0_16:1)    NA     NA     NA            NA           NA    NA     NA
#>                    FA2 DB FA2 Bond Type LCB Position LCB C LCB OH LCB DB
#> PC 32:1                NA            NA           NA    NA     NA     NA
#> LPC 34:1               NA            NA           NA    NA     NA     NA
#> TG(18:1_18:0_16:1)     NA            NA           NA    NA     NA     NA
#>                    LCB Bond Type FA3 Position FA3 C FA3 OH FA3 DB FA3 Bond Type
#> PC 32:1                       NA           NA    NA     NA     NA            NA
#> LPC 34:1                      NA           NA    NA     NA     NA            NA
#> TG(18:1_18:0_16:1)            NA           NA    NA     NA     NA            NA
#>                    FA4 Position FA4 C FA4 OH FA4 DB FA4 Bond Type
#> PC 32:1                      NA    NA     NA     NA            NA
#> LPC 34:1                     NA    NA     NA     NA            NA
#> TG(18:1_18:0_16:1)           NA    NA     NA     NA            NA