Parse lipid name with a specific grammar.

parseLipidName reads the provided lipid name based on grammar and returns structural information as a named vector. Will return a vector with the "Grammar" element set to "NOT_PARSEABLE" if none of the parsers was able to parse the provided name with the provided grammar successfully. If any error was raised, returns an empty vector.

parseLipidNameWithGrammar(lipidName, grammar)

Arguments

lipidName
The lipid name to parse.

grammar
The grammar to use. One of "Goslin", "GoslinFragments", "SwissLipids", "LipidMaps", "HMDB"

Value

Named vector with details of the lipid, empty vector otherwise.

Examples

parseLipidNameWithGrammar("PC 32:1", "Goslin")

#>                 Normalized Name                   Original Name 
#>                       "PC 32:1"                       "PC 32:1" 
#>                         Grammar             Lipid Maps Category 
#>                        "Goslin"                            "GP" 
#>           Lipid Maps Main Class                    Species Name 
#> "Glycerophosphocholines [GP01]"                       "PC 32:1" 
#>       Molecular Subspecies Name      Structural Subspecies Name 
#>                            "NA"                            "NA" 
#>        Isomeric Subspecies Name           Functional Class Abbr 
#>                            "NA"                          "[PC]" 
#>       Functional Class Synonyms                           Level 
#>                          "[PC]"                       "SPECIES" 
#>                         Total C                        Total OH 
#>                            "32"                             "0" 
#>                        Total DB                            Mass 
#>                             "1"                       "731.547" 
#>                     Sum Formula                    FA1 Position 
#>                    "C40H78NO8P"                            "NA" 
#>                           FA1 C                          FA1 OH 
#>                            "NA"                            "NA" 
#>                          FA1 DB                   FA1 Bond Type 
#>                            "NA"                            "NA" 
#>                FA1 DB Positions                    FA2 Position 
#>                            "NA"                            "NA" 
#>                           FA2 C                          FA2 OH 
#>                            "NA"                            "NA" 
#>                          FA2 DB                   FA2 Bond Type 
#>                            "NA"                            "NA" 
#>                FA2 DB Positions                    LCB Position 
#>                            "NA"                            "NA" 
#>                           LCB C                          LCB OH 
#>                            "NA"                            "NA" 
#>                          LCB DB                   LCB Bond Type 
#>                            "NA"                            "NA" 
#>                    FA3 Position                           FA3 C 
#>                            "NA"                            "NA" 
#>                          FA3 OH                          FA3 DB 
#>                            "NA"                            "NA" 
#>                   FA3 Bond Type                FA3 DB Positions 
#>                            "NA"                            "NA" 
#>                    FA4 Position                           FA4 C 
#>                            "NA"                            "NA" 
#>                          FA4 OH                          FA4 DB 
#>                            "NA"                            "NA" 
#>                   FA4 Bond Type                FA4 DB Positions 
#>                            "NA"                            "NA" 

parseLipidNameWithGrammar("LPC(34:1)", "SwissLipids")

#>                           Normalized Name 
#>                                "LPC 34:1" 
#>                             Original Name 
#>                               "LPC(34:1)" 
#>                                   Grammar 
#>                             "SwissLipids" 
#>                       Lipid Maps Category 
#>                                      "GP" 
#>                     Lipid Maps Main Class 
#> "Monoacylglycerophosphocholines [GP0105]" 
#>                              Species Name 
#>                                "LPC 34:1" 
#>                 Molecular Subspecies Name 
#>                                      "NA" 
#>                Structural Subspecies Name 
#>                                      "NA" 
#>                  Isomeric Subspecies Name 
#>                                      "NA" 
#>                     Functional Class Abbr 
#>                                   "[LPC]" 
#>                 Functional Class Synonyms 
#>                           "[LPC, LysoPC]" 
#>                                     Level 
#>                                 "SPECIES" 
#>                                   Total C 
#>                                      "34" 
#>                                  Total OH 
#>                                       "0" 
#>                                  Total DB 
#>                                       "1" 
#>                                      Mass 
#>                                 "745.599" 
#>                               Sum Formula 
#>                              "C42H84NO7P" 
#>                              FA1 Position 
#>                                      "NA" 
#>                                     FA1 C 
#>                                      "NA" 
#>                                    FA1 OH 
#>                                      "NA" 
#>                                    FA1 DB 
#>                                      "NA" 
#>                             FA1 Bond Type 
#>                                      "NA" 
#>                          FA1 DB Positions 
#>                                      "NA" 
#>                              FA2 Position 
#>                                      "NA" 
#>                                     FA2 C 
#>                                      "NA" 
#>                                    FA2 OH 
#>                                      "NA" 
#>                                    FA2 DB 
#>                                      "NA" 
#>                             FA2 Bond Type 
#>                                      "NA" 
#>                          FA2 DB Positions 
#>                                      "NA" 
#>                              LCB Position 
#>                                      "NA" 
#>                                     LCB C 
#>                                      "NA" 
#>                                    LCB OH 
#>                                      "NA" 
#>                                    LCB DB 
#>                                      "NA" 
#>                             LCB Bond Type 
#>                                      "NA" 
#>                              FA3 Position 
#>                                      "NA" 
#>                                     FA3 C 
#>                                      "NA" 
#>                                    FA3 OH 
#>                                      "NA" 
#>                                    FA3 DB 
#>                                      "NA" 
#>                             FA3 Bond Type 
#>                                      "NA" 
#>                          FA3 DB Positions 
#>                                      "NA" 
#>                              FA4 Position 
#>                                      "NA" 
#>                                     FA4 C 
#>                                      "NA" 
#>                                    FA4 OH 
#>                                      "NA" 
#>                                    FA4 DB 
#>                                      "NA" 
#>                             FA4 Bond Type 
#>                                      "NA" 
#>                          FA4 DB Positions 
#>                                      "NA" 

parseLipidNameWithGrammar("TG(18:1_18:0_16:1)", "LipidMaps")

#>             Normalized Name               Original Name 
#>        "TAG 18:1-18:0-16:1"        "TG(18:1_18:0_16:1)" 
#>                     Grammar         Lipid Maps Category 
#>                 "LipidMaps"                        "GL" 
#>       Lipid Maps Main Class                Species Name 
#> "Triacylglycerols [GL0301]"                  "TAG 52:2" 
#>   Molecular Subspecies Name  Structural Subspecies Name 
#>        "TAG 18:1-18:0-16:1"                        "NA" 
#>    Isomeric Subspecies Name       Functional Class Abbr 
#>                        "NA"                      "[TG]" 
#>   Functional Class Synonyms                       Level 
#>                 "[TAG, TG]"      "MOLECULAR_SUBSPECIES" 
#>                     Total C                    Total OH 
#>                        "52"                         "0" 
#>                    Total DB                        Mass 
#>                         "2"                   "858.768" 
#>                 Sum Formula                FA1 Position 
#>                 "C55H102O6"                         "0" 
#>                       FA1 C                      FA1 OH 
#>                        "18"                         "0" 
#>                      FA1 DB               FA1 Bond Type 
#>                         "1"                     "ESTER" 
#>            FA1 DB Positions                FA2 Position 
#>                        "[]"                         "0" 
#>                       FA2 C                      FA2 OH 
#>                        "18"                         "0" 
#>                      FA2 DB               FA2 Bond Type 
#>                         "0"                     "ESTER" 
#>            FA2 DB Positions                LCB Position 
#>                        "[]"                        "NA" 
#>                       LCB C                      LCB OH 
#>                        "NA"                        "NA" 
#>                      LCB DB               LCB Bond Type 
#>                        "NA"                        "NA" 
#>                FA3 Position                       FA3 C 
#>                         "0"                        "16" 
#>                      FA3 OH                      FA3 DB 
#>                         "0"                         "1" 
#>               FA3 Bond Type            FA3 DB Positions 
#>                     "ESTER"                        "[]" 
#>                FA4 Position                       FA4 C 
#>                        "NA"                        "NA" 
#>                      FA4 OH                      FA4 DB 
#>                        "NA"                        "NA" 
#>               FA4 Bond Type            FA4 DB Positions 
#>                        "NA"                        "NA"